+++ src/ariadne_navitree.h
@@ -33,22 +33,22 @@ extern "C" {
 /******************************************************************************/
 #pragma pack(1)
 
-extern long tree_val[TREEVALS_COUNT];
+extern int32_t tree_val[TREEVALS_COUNT];
 
 #pragma pack()
 /******************************************************************************/
 void tags_init(void);
-long update_border_tags(long tag_id, int32_t *border_pt, long border_len);
-long border_tags_to_current(int32_t *border_pt, long border_len);
-TbBool is_current_tag(long tag_id);
-void store_current_tag(long tag_id);
+int32_t update_border_tags(int32_t tag_id, int32_t *border_pt, int32_t border_len);
+int32_t border_tags_to_current(int32_t *border_pt, int32_t border_len);
+TbBool is_current_tag(int32_t tag_id);
+void store_current_tag(int32_t tag_id);
 
-TbBool navitree_add(long itm_pos, long itm_dat, long mvcost);
-long copy_tree_to_route(long tag_start_id, long tag_end_id, int32_t *route_pts, long route_len);
+TbBool navitree_add(int32_t itm_pos, int32_t itm_dat, int32_t mvcost);
+int32_t copy_tree_to_route(int32_t tag_start_id, int32_t tag_end_id, int32_t *route_pts, int32_t route_len);
 
 void delaunay_init(void);
-TbBool delaunay_add(long itm_pos);
-long delaunay_seeded(long start_x, long start_y, long end_x, long end_y, TbBool keep_edge);
+TbBool delaunay_add(int32_t itm_pos);
+int32_t delaunay_seeded(int32_t start_x, int32_t start_y, int32_t end_x, int32_t end_y, TbBool keep_edge);
 /******************************************************************************/
 #ifdef __cplusplus
 }
