+++ src/ariadne_tringls.c
@@ -35,30 +35,30 @@ extern "C" {
 #endif
 /******************************************************************************/
 
-long free_Triangles = 0;
+int32_t free_Triangles = 0;
 struct Triangle Triangles[TRIANLGLES_COUNT];
-long count_Triangles = 0;
-long ix_Triangles = 0;
+int32_t count_Triangles = 0;
+int32_t ix_Triangles = 0;
 
 /******************************************************************************/
 struct Triangle bad_triangle;
-const long MOD3[] = {0, 1, 2, 0, 1, 2};
+const int32_t MOD3[] = {0, 1, 2, 0, 1, 2};
 /******************************************************************************/
-long tri_new(void)
+int32_t tri_new(void)
 {
-    long i;
+    int32_t i;
     if (free_Triangles == -1)
     {
         i = ix_Triangles;
         if ((i < 0) || (i >= TRIANLGLES_COUNT))
         {
-            ERRORLOG("ix_Triangles overflow, got %ld",i);
+            ERRORLOG("ix_Triangles overflow, got %d",i);
             erstat_inc(ESE_NoFreeTriangls);
             return -1;
         }
         if (i > (TRIANLGLES_COUNT * 999 / 1000))
         {
-            WARNLOG("TRIANGLES near limit: %ld", i);
+            WARNLOG("TRIANGLES near limit: %d", i);
         }
         ix_Triangles++;
     } else
@@ -66,7 +66,7 @@ long tri_new(void)
         i = free_Triangles;
         if ((i < 0) || (i >= TRIANLGLES_COUNT))
         {
-            ERRORLOG("free_Triangles overflow, got %ld",i);
+            ERRORLOG("free_Triangles overflow, got %d",i);
             erstat_inc(ESE_NoFreeTriangls);
             return -1;
         }
@@ -77,23 +77,23 @@ long tri_new(void)
     return i;
 }
 
-void tri_dispose(long tri_idx)
+void tri_dispose(int32_t tri_idx)
 {
-    long pfree_idx = free_Triangles;
+    int32_t pfree_idx = free_Triangles;
     free_Triangles = tri_idx;
     Triangles[tri_idx].tags[0] = pfree_idx;
     Triangles[tri_idx].tree_alt = NAV_COL_UNSET;
     count_Triangles--;
 }
 
-long get_triangle_region_id(long tri_id)
+int32_t get_triangle_region_id(int32_t tri_id)
 {
     if ((tri_id < 0) || (tri_id >= TRIANLGLES_COUNT))
         return -1;
     return (Triangles[tri_id].region_and_edgelen >> EDGELEN_BITS);
 }
 
-TbBool set_triangle_region_id(long tri_id, long reg_id)
+TbBool set_triangle_region_id(int32_t tri_id, int32_t reg_id)
 {
     if ((tri_id < 0) || (tri_id >= TRIANLGLES_COUNT))
         return false;
@@ -102,14 +102,14 @@ TbBool set_triangle_region_id(long tri_id, long reg_id)
     return true;
 }
 
-long get_triangle_edgelen(long tri_id)
+int32_t get_triangle_edgelen(int32_t tri_id)
 {
     if ((tri_id < 0) || (tri_id >= TRIANLGLES_COUNT))
         return 0;
     return (Triangles[tri_id].region_and_edgelen & ((1 << EDGELEN_BITS) - 1));
 }
 
-TbBool set_triangle_edgelen(long tri_id, long edgelen)
+TbBool set_triangle_edgelen(int32_t tri_id, int32_t edgelen)
 {
     if ((tri_id < 0) || (tri_id >= TRIANLGLES_COUNT))
         return false;
@@ -118,14 +118,14 @@ TbBool set_triangle_edgelen(long tri_id, long edgelen)
     return true;
 }
 
-NavColour get_triangle_tree_alt(long tri_id)
+NavColour get_triangle_tree_alt(int32_t tri_id)
 {
     if ((tri_id < 0) || (tri_id >= TRIANLGLES_COUNT))
         return -1;
     return Triangles[tri_id].tree_alt;
 }
 
-struct Triangle *get_triangle(long tri_id)
+struct Triangle *get_triangle(int32_t tri_id)
 {
     if ((tri_id < 0) || (tri_id >= TRIANLGLES_COUNT))
         return INVALID_TRIANGLE;
@@ -137,7 +137,7 @@ TbBool triangle_is_invalid(const struct Triangle *tri)
     return (tri < &Triangles[0]) || (tri > &Triangles[TRIANLGLES_COUNT-1]) || (tri == INVALID_TRIANGLE) || (tri == NULL);
 }
 
-struct Point *get_triangle_point(long tri_id, long pt_cor)
+struct Point *get_triangle_point(int32_t tri_id, int32_t pt_cor)
 {
     if ((tri_id < 0) || (tri_id >= TRIANLGLES_COUNT))
         return INVALID_POINT;
@@ -146,23 +146,23 @@ struct Point *get_triangle_point(long tri_id, long pt_cor)
     return point_get(Triangles[tri_id].points[pt_cor]);
 }
 
-TbBool triangle_tip_equals(long tri_id, long pt_cor, long pt_x, long pt_y)
+TbBool triangle_tip_equals(int32_t tri_id, int32_t pt_cor, int32_t pt_x, int32_t pt_y)
 {
     if ((tri_id < 0) || (tri_id >= TRIANLGLES_COUNT))
         return false;
     if ((pt_cor < 0) || (pt_cor >= 3))
         return false;
-    long pt_id = Triangles[tri_id].points[pt_cor];
+    int32_t pt_id = Triangles[tri_id].points[pt_cor];
     return point_equals(pt_id, pt_x, pt_y);
 }
 
-long link_find(long ntri, long val)
+int32_t link_find(int32_t ntri, int32_t val)
 {
     if ((ntri < 0) || (ntri >= TRIANLGLES_COUNT))
     {
         return -1;
     }
-    for (long i = 0; i < 3; i++)
+    for (int32_t i = 0; i < 3; i++)
     {
         if (Triangles[ntri].tags[i] == val)
         {
@@ -172,10 +172,10 @@ long link_find(long ntri, long val)
     return -1;
 }
 
-TbBool outer_locked(long ntri, long ncor)
+TbBool outer_locked(int32_t ntri, int32_t ncor)
 {
-    long n = Triangles[ntri].tags[ncor];
-    long shft = link_find(n, ntri);
+    int32_t n = Triangles[ntri].tags[ncor];
+    int32_t shft = link_find(n, ntri);
     if (shft < 0)
     {
         ERRORLOG("border edge");
@@ -184,18 +184,18 @@ TbBool outer_locked(long ntri, long ncor)
     return ( (Triangles[n].navigation_flags & (1 << (shft + 3)) ) != 0);
 }
 
-long point_loop(long pt_tri, long pt_cor)
+int32_t point_loop(int32_t pt_tri, int32_t pt_cor)
 {
-    long ntri = pt_tri;
-    long ncor = pt_cor;
+    int32_t ntri = pt_tri;
+    int32_t ncor = pt_cor;
     if (ntri < 0)
         return -1;
-    long k = 0;
-    long n;
+    int32_t k = 0;
+    int32_t n;
     do
     {
         n = Triangles[ntri].tags[ncor];
-        long i = link_find(n, ntri);
+        int32_t i = link_find(n, ntri);
         if (i < 0)
             return -1;
         ncor = MOD3[i+1];
@@ -207,10 +207,10 @@ long point_loop(long pt_tri, long pt_cor)
     return k;
 }
 
-void edgelen_set(long tri_id)
+void edgelen_set(int32_t tri_id)
 {
     NAVIDBG(19,"Starting");
-    static const unsigned long EdgeLenBits[][4] = {
+    static const uint32_t EdgeLenBits[][4] = {
         {1, 1, 2, 3},
         {1, 1, 2, 3},
         {2, 2, 2, 3},
@@ -225,7 +225,7 @@ void edgelen_set(long tri_id)
         delta_x = 3;
     if (delta_y > 3)
         delta_y = 3;
-    unsigned long edge_len = (EdgeLenBits[delta_y][delta_x] << 4);
+    uint32_t edge_len = (EdgeLenBits[delta_y][delta_x] << 4);
     int pt1idx = tri->points[1];
     delta_x = abs(ari_Points[pt1idx].x - ari_Points[pt2idx].x);
     delta_y = abs(ari_Points[pt1idx].y - ari_Points[pt2idx].y);
@@ -244,9 +244,9 @@ void edgelen_set(long tri_id)
     set_triangle_edgelen(tri_id, edge_len);
 }
 
-long edge_rotateAC(long tri1_id, long cor1_id)
+int32_t edge_rotateAC(int32_t tri1_id, int32_t cor1_id)
 {
-    long tri2_id = Triangles[tri1_id].tags[cor1_id];
+    int32_t tri2_id = Triangles[tri1_id].tags[cor1_id];
     if (tri2_id == -1) {
         return false;
     }
@@ -260,8 +260,8 @@ long edge_rotateAC(long tri1_id, long cor1_id)
     int cor1c_id = MOD3[cor1_id + 2];
     int cor2c_id = MOD3[cor2_id + 2];
 
-    long tri3_id = Triangles[tri1_id].tags[cor1b_id];
-    long tri4_id = Triangles[tri2_id].tags[cor2b_id];
+    int32_t tri3_id = Triangles[tri1_id].tags[cor1b_id];
+    int32_t tri4_id = Triangles[tri2_id].tags[cor2b_id];
 
     {
         unsigned short tri1_fld = Triangles[tri1_id].navigation_flags;
@@ -279,17 +279,17 @@ long edge_rotateAC(long tri1_id, long cor1_id)
 
     int pt1_id = Triangles[tri1_id].points[cor1_id];
     int pt2_id = Triangles[tri1_id].points[cor1c_id];
-    long diff_ax = ari_Points[pt1_id].x - ari_Points[pt2_id].x;
-    long diff_ay = ari_Points[pt1_id].y - ari_Points[pt2_id].y;
+    int32_t diff_ax = ari_Points[pt1_id].x - ari_Points[pt2_id].x;
+    int32_t diff_ay = ari_Points[pt1_id].y - ari_Points[pt2_id].y;
     int pt3_id = Triangles[tri2_id].points[cor2c_id];
-    long diff_bx = ari_Points[pt3_id].x - ari_Points[pt2_id].x;
-    long diff_by = ari_Points[pt3_id].y - ari_Points[pt2_id].y;
+    int32_t diff_bx = ari_Points[pt3_id].x - ari_Points[pt2_id].x;
+    int32_t diff_by = ari_Points[pt3_id].y - ari_Points[pt2_id].y;
     if (LbCompareMultiplications(diff_ay, diff_bx, diff_ax, diff_by) <= 0) {
         return false;
     }
     int pt4_id = Triangles[tri2_id].points[cor2_id];
-    long diff_cx = ari_Points[pt4_id].x - ari_Points[pt2_id].x;
-    long diff_cy = ari_Points[pt4_id].y - ari_Points[pt2_id].y;
+    int32_t diff_cx = ari_Points[pt4_id].x - ari_Points[pt2_id].x;
+    int32_t diff_cy = ari_Points[pt4_id].y - ari_Points[pt2_id].y;
     if (LbCompareMultiplications(diff_cy, diff_bx, diff_cx, diff_by) >= 0) {
         return false;
     }
@@ -329,13 +329,13 @@ long edge_rotateAC(long tri1_id, long cor1_id)
     edgelen_set(tri2_id);
     return true;
 }
-long reduce_point(int32_t *pt_tri, int32_t *pt_cor)
+int32_t reduce_point(int32_t *pt_tri, int32_t *pt_cor)
 {
-    long k = 0;
-    long ntri = *pt_tri;
-    long ncor = *pt_cor;
-    long first_tri = *pt_tri;
-    long ctri;
+    int32_t k = 0;
+    int32_t ntri = *pt_tri;
+    int32_t ncor = *pt_cor;
+    int32_t first_tri = *pt_tri;
+    int32_t ctri;
     do
     {
         ctri = Triangles[ntri].tags[ncor];
@@ -343,7 +343,7 @@ long reduce_point(int32_t *pt_tri, int32_t *pt_cor)
             return -1;
         if (!edge_rotateAC(ntri, ncor))
         {
-            long i = link_find(ctri, ntri);
+            int32_t i = link_find(ctri, ntri);
             if (i < 0)
                 return -1;
             ncor = MOD3[i+1];
@@ -357,9 +357,9 @@ long reduce_point(int32_t *pt_tri, int32_t *pt_cor)
     return k;
 }
 
-long triangle_find_first_used(void)
+int32_t triangle_find_first_used(void)
 {
-    for (long tri_idx = 0; tri_idx < ix_Triangles; tri_idx++)
+    for (int32_t tri_idx = 0; tri_idx < ix_Triangles; tri_idx++)
     {
         struct Triangle* tri = &Triangles[tri_idx];
         if (tri->tree_alt != NAV_COL_UNSET) {
@@ -370,7 +370,7 @@ long triangle_find_first_used(void)
     return -1;
 }
 
-void triangulation_init_triangles(long pt_id1, long pt_id2, long pt_id3, long pt_id4)
+void triangulation_init_triangles(int32_t pt_id1, int32_t pt_id2, int32_t pt_id3, int32_t pt_id4)
 {
     free_Triangles = -1;
     ix_Triangles = 2;
@@ -395,14 +395,14 @@ void triangulation_init_triangles(long pt_id1, long pt_id2, long pt_id3, long pt
     Triangles[1].navigation_flags = 7;
 }
 
-char triangle_divide_areas_s8differ(long ntri, long ncorA, long ncorB, long pt_x, long pt_y)
+char triangle_divide_areas_s8differ(int32_t ntri, int32_t ncorA, int32_t ncorB, int32_t pt_x, int32_t pt_y)
 {
     struct Point* pt = get_triangle_point(ntri, ncorA);
-    long tipA_x = (pt->x << 8);
-    long tipA_y = (pt->y << 8);
+    int32_t tipA_x = (pt->x << 8);
+    int32_t tipA_y = (pt->y << 8);
     pt = get_triangle_point(ntri,ncorB);
-    long tipB_x = (pt->x << 8);
-    long tipB_y = (pt->y << 8);
+    int32_t tipB_x = (pt->x << 8);
+    int32_t tipB_y = (pt->y << 8);
     return LbCompareMultiplications(tipA_x-pt_x, tipB_y-pt_y, tipA_y-pt_y, tipB_x-pt_x);
 }
 /******************************************************************************/
